pyproject.toml
[project]
name = "database-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "psycopg2-binary",
    "python-dateutil",
]
